split project into 'core' and 'paper' module
