+++ build.gradle
@@ -3,40 +3,37 @@ plugins {
     id 'maven-publish'
 }
 
-group = 'grabsky.configuration'
-version = '0.9.4-pre'
+subprojects { subproject ->
+    apply plugin: 'java-library'
+    apply plugin: 'maven-publish'
 
-sourceCompatibility = JavaVersion.VERSION_17
-targetCompatibility = JavaVersion.VERSION_17
+    group = 'grabsky.configuration'
+    version = '0.9.4-pre'
 
-publishing {
-    publications {
-        maven(MavenPublication) {
-            groupId = 'com.github.grabsky'
-            artifactId = rootProject.name
-            from components.java
+    sourceCompatibility = JavaVersion.VERSION_17
+    targetCompatibility = JavaVersion.VERSION_17
+
+    publishing {
+        publications {
+            maven(MavenPublication) {
+                groupId = 'com.github.grabsky'
+                artifactId = rootProject.name + '-' + subproject.name
+                from components.java
+            }
         }
     }
-}
 
-repositories {
-    mavenCentral()
-    maven { url = 'https://papermc.io/repo/repository/maven-public/' }
-}
+    repositories {
+        mavenCentral()
+    }
 
-dependencies {
-    testImplementation 'org.junit.jupiter:junit-jupiter-api:5.8.1'
-    testRuntimeOnly 'org.junit.jupiter:junit-jupiter-engine:5.8.1'
-    // Lombok
-    compileOnly 'org.projectlombok:lombok:1.18.24'
-    annotationProcessor 'org.projectlombok:lombok:1.18.24'
-    // Paper API
-    compileOnly 'io.papermc.paper:paper-api:1.19.2-R0.1-SNAPSHOT'
-    // HOCON
-    compileOnly 'com.typesafe:config:1.4.2'
-    testImplementation 'com.typesafe:config:1.4.2'
-}
+    dependencies {
+        // Lombok
+        compileOnly 'org.projectlombok:lombok:1.18.24'
+        annotationProcessor 'org.projectlombok:lombok:1.18.24'
 
-test {
-    useJUnitPlatform()
+        // JetBrains Annotations
+        compileOnly("org.jetbrains:annotations:23.0.0")
+    }
+
 }
