+++ core/build.gradle
@@ -1,0 +1,12 @@
+test {
+    useJUnitPlatform()
+}
+
+dependencies {
+    // JUnit 5
+    testImplementation 'org.junit.jupiter:junit-jupiter-api:5.8.1'
+    testRuntimeOnly 'org.junit.jupiter:junit-jupiter-engine:5.8.1'
+    // lightbend/config
+    compileOnly 'com.typesafe:config:1.4.2'
+    testImplementation 'com.typesafe:config:1.4.2'
+}
