+++ core/src/main/java/grabsky/configuration/exceptions/SerializationException.java
