+++ core/src/main/java/grabsky/configuration/serialization/TypeSerializer.java
