+++ core/src/main/java/grabsky/configuration/util/Numbers.java
