+++ core/src/main/java/grabsky/configuration/util/SerializerContainer.java
