+++ core/src/test/resources/test.conf
