+++ paper/build.gradle
@@ -1,0 +1,12 @@
+repositories {
+    mavenCentral()
+}
+
+repositories {
+    maven { url = 'https://repo.papermc.io/repository/maven-public/' }
+}
+
+dependencies {
+    compileOnly project(':core')
+    compileOnly 'io.papermc.paper:paper-api:1.19.2-R0.1-SNAPSHOT'
+}
