+++ settings.gradle
@@ -1,1 +1,5 @@
 rootProject.name = 'configuration'
+
+include 'core'
+include 'paper'
+
